+++ src/drx-sdk-dotnet/Receipt/LineItem/General/Book.cs
@@ -21,14 +21,14 @@ namespace Net.Dreceiptx.Receipt.LineItem.General
         public static readonly string LineItemTypeValue = "GENERAL0001";
         private readonly string _tradeItemIdentificationISBN = "ISBN";
 
-        public Book(TradeItemDescriptionInformation tradeItemDescriptionInformation, int quantity, double price) 
+        public Book(TradeItemDescriptionInformation tradeItemDescriptionInformation, int quantity, decimal price) 
             : base(tradeItemDescriptionInformation, quantity, price)
         {
             _transactionalTradeItemType = TransactionalTradeItemType.MANUAL;
             AddTradeItemIdentification(LineItem.LineItemTypeIdentifier, Book.LineItemTypeValue);
         }
 
-        public Book(string ISBN, int quantity, double price) :base()
+        public Book(string ISBN, int quantity, decimal price) :base()
         {
             AddTradeItemIdentification(LineItem.LineItemTypeIdentifier, Book.LineItemTypeValue);
             AddTradeItemIdentification(_tradeItemIdentificationISBN, ISBN);
